GasLiftSingleWell: split typetag independent code to separate class
