+++ CMakeLists_files.cmake
@@ -53,6 +53,7 @@ list (APPEND MAIN_SOURCE_FILES
   opm/simulators/utils/ParallelFileMerger.cpp
   opm/simulators/utils/ParallelRestart.cpp
   opm/simulators/wells/ALQState.cpp
+  opm/simulators/wells/GasLiftSingleWellGeneric.cpp
   opm/simulators/wells/GlobalWellInfo.cpp
   opm/simulators/wells/GroupState.cpp
   opm/simulators/wells/ParallelWellInfo.cpp
